Make as a plugin
